requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
discord.py==1.7.3
idna==3.2
importlib-metadata==4.7.1
joblib==1.0.1
multidict==5.1.0
nltk==3.6.2
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.27
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
tqdm==4.62.2
typing-extensions==3.10.0.0
yarl==1.6.3
zipp==3.5.0
